+++ hphp/runtime/ext/process/ext_process.h
@@ -77,47 +77,6 @@ int64_t HHVM_FUNCTION(pcntl_wtermsig,
                       int status);
 
 ///////////////////////////////////////////////////////////////////////////////
-
-Variant HHVM_FUNCTION(shell_exec,
-                      const String& cmd);
-String HHVM_FUNCTION(exec,
-                     const String& command,
-                     VRefParam output = uninit_null(),
-                     VRefParam return_var = uninit_null());
-void HHVM_FUNCTION(passthru,
-                   const String& command,
-                   VRefParam return_var = uninit_null());
-String HHVM_FUNCTION(system,
-                     const String& command,
-                     VRefParam return_var = uninit_null());
-
-///////////////////////////////////////////////////////////////////////////////
-
-Variant HHVM_FUNCTION(proc_open,
-                      const String& cmd,
-                      const Array& descriptorspec,
-                      VRefParam pipes,
-                      const String& cwd = null_string,
-                      const Variant& env = null_variant,
-                      const Variant& other_options = null_variant);
-bool HHVM_FUNCTION(proc_terminate,
-                   const Resource& process,
-                   int signal = SIGTERM);
-int64_t HHVM_FUNCTION(proc_close,
-                      const Resource& process);
-Array HHVM_FUNCTION(proc_get_status,
-                    const Resource& process);
-bool HHVM_FUNCTION(proc_nice,
-                   int increment);
-
-///////////////////////////////////////////////////////////////////////////////
-
-String HHVM_FUNCTION(escapeshellarg,
-                     const String& arg);
-String HHVM_FUNCTION(escapeshellcmd,
-                     const String& command);
-
-///////////////////////////////////////////////////////////////////////////////
 }
 
 #endif // incl_HPHP_EXT_PROCESS_H_
